requirements.txt
Django>=5.0
djangorestframework
fastapi
uvicorn
python-dotenv
supabase
psycopg2-binary
requests
Pillow